% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crowdr.R
\name{scrape_reviews}
\alias{scrape_reviews}
\title{Scrape review texts from urls}
\usage{
scrape_reviews(urls)
}
\arguments{
\item{urls}{Vector of URLs}
}
\description{
Summarizes merging path by giving pairs of factor groups merged
in each iteration.
}
\examples{
x <- scrape_revies(urls)
}
