% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crowdr.R
\name{scrape_urls}
\alias{scrape_urls}
\title{Scrape URLs}
\usage{
scrape_urls(category, pages = NULL)
}
\arguments{
\item{category}{String. Slug of a software category}

\item{pages}{Integer. Optional. Number of pages that should be scraped.
By default \code{pages} will scrape all available pages.}
}
\value{
Returns a `tibble` with name, URL, number of ratings, and average
        rating for each tool of a software category.
}
\description{
Scrape tool names, URLs, and ratings meta data from a software category on
'G2Crowd.com'.
}
\examples{
x <- scrape_urls("crm", 3)
}
