% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crowdr.R
\name{scrape_reviews}
\alias{scrape_reviews}
\title{Scrape review texts from urls}
\usage{
scrape_reviews(urls)
}
\arguments{
\item{factorMerger}{Object of a class \code{factorMerger}}

\item{showStats}{If \code{TRUE} extends results with
the loglikelihood (column \code{model}),
p-value for the \code{LRT} tests against the full model (column \code{pval})
and Generalized Information Criterion value (column \code{GIC}).
By default \code{showStats} is set to \code{FALSE}.}

\item{round}{Logical. If \code{TRUE}, the default, statistics are rounded}

\item{penalty}{GIC penalty}
}
\description{
Summarizes merging path by giving pairs of factor groups merged
in each iteration.
}
\examples{
randSample <- generateMultivariateSample(N = 100, k = 10, d = 3)
fm <- mergeFactors(randSample$response, randSample$factor)
mergingHistory(fm, showStats = TRUE)
}
